LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/element-request-fullscreen-namespaces-expected.txt


FAIL requestFullscreen() fails for an element in null namespace assert_unreached: fullscreenchange event Reached unreachable code
FAIL requestFullscreen() fails for an element in https://unknown.namespace namespace promise_rejects_js: function "function () { throw e }" threw object "Error: element click intercepted error" ("Error") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL requestFullscreen() fails for an element in  namespace promise_rejects_js: function "function () { throw e }" threw object "Error: element click intercepted error" ("Error") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL requestFullscreen() succeed for an element in http://www.w3.org/1999/xhtml namespace promise_test: Unhandled rejection with value: object "Error: element click intercepted error"
FAIL requestFullscreen() succeed for an element in http://www.w3.org/2000/svg namespace promise_test: Unhandled rejection with value: object "Error: element click intercepted error"
FAIL requestFullscreen() succeed for an element in http://www.w3.org/1998/Math/MathML namespace promise_test: Unhandled rejection with value: object "Error: element click intercepted error"
click to continue test
click to continue test
click to continue test
click to continue test
click to continue test
